Ignore Pipfile and Pipfile.lock in the Git version control
The .gitignore file has been updated to include Pipfile and
Pipfile.lock.
Ignoring these files prevents them from being tracked in the version
control system.
This change is beneficial because Pipfile and Pipfile.lock are usually
not intended to be versioned and should not be included in the project's
repository.
No issues are associated with this change.